+++ src/schema/build-context.ts
@@ -7,8 +7,6 @@ import { Middleware } from "../interfaces/Middleware";
 import { ContainerType, ContainerGetter, IOCContainer } from "../utils/container";
 import { ValidatorFn } from "../interfaces/ValidatorFn";
 
-export type DateScalarMode = "isoDate" | "timestamp";
-
 export interface ScalarsTypeMap {
   type: Function;
   scalar: GraphQLScalarType;
@@ -17,7 +15,6 @@ export interface ScalarsTypeMap {
 export type ValidateSettings = boolean | ValidatorOptions | ValidatorFn<object>;
 
 export interface BuildContextOptions {
-  dateScalarMode?: DateScalarMode;
   scalarsMap?: ScalarsTypeMap[];
   /**
    * Indicates if class-validator should be used to auto validate objects injected into params.
@@ -41,7 +38,6 @@ export interface BuildContextOptions {
 }
 
 export abstract class BuildContext {
-  static dateScalarMode: DateScalarMode;
   static scalarsMaps: ScalarsTypeMap[];
   static validate: ValidateSettings;
   static authChecker?: AuthChecker<any, any>;
@@ -56,10 +52,6 @@ export abstract class BuildContext {
    * Set static fields with current building context data
    */
   static create(options: BuildContextOptions) {
-    if (options.dateScalarMode !== undefined) {
-      this.dateScalarMode = options.dateScalarMode;
-    }
-
     if (options.scalarsMap !== undefined) {
       this.scalarsMaps = options.scalarsMap;
     }
@@ -103,7 +95,6 @@ export abstract class BuildContext {
    * Restore default settings
    */
   static reset() {
-    this.dateScalarMode = "isoDate";
     this.scalarsMaps = [];
     this.validate = true;
     this.authChecker = undefined;